requirements.txt
Django==2.2
djangorestframework==3.9.2
nltk==3.4
numpy==1.16.2
scikit-learn==0.20.3
pandas==0.24.2
beautifulsoup4==4.7.1
lxml==4.3.3
django-cors-headers==2.5.2